requirements.txt
discord
qrcode
web3
Image
pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client